Use only agg accuracy
